Hide applies to column for files
